requirements.txt
fastapi
uvicorn
sentence-transformers
transformers
torch>=2.4
nltk
scikit-learn
pdfplumber
python-docx